Migrate to DI
