+++ .cuprc.js
@@ -1,6 +1,5 @@
 module.exports = {
   babel: {
-    plugins: [require.resolve('@babel/plugin-transform-flow-strip-types')],
     presets: [require.resolve('@babel/preset-react')],
   },
 };
