+++ package.json
@@ -4,54 +4,53 @@
   "repository": "fusionjs/fusion-plugin-csrf-protection",
   "version": "0.2.2",
   "files": [
-    "dist"
+    "dist",
+    "src"
   ],
-  "main": "./dist/node.cjs.js",
-  "module": "./dist/node.es.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.es.js",
   "browser": {
-    "./dist/node.cjs.js": "./dist/browser.es5.cjs.js",
-    "./dist/node.es.js": "./dist/browser.es5.es.js"
+    "./dist/index.js": "./dist/browser.es5.js",
+    "./dist/index.es.js": "./dist/browser.es5.es.js"
   },
   "es2015": {
-    "./dist/browser.es5.cjs.js": "./dist/browser.es2015.cjs.js",
     "./dist/browser.es5.es.js": "./dist/browser.es2015.es.js"
   },
   "es2017": {
-    "./dist/browser.es5.cjs.js": "./dist/browser.es2017.cjs.js",
     "./dist/browser.es5.es.js": "./dist/browser.es2017.es.js",
-    "./dist/browser.es2015.cjs.js": "./dist/browser.es2017.cjs.js",
     "./dist/browser.es2015.es.js": "./dist/browser.es2017.es.js"
   },
   "dependencies": {
     "base64-url": "^2.0.1"
   },
   "devDependencies": {
-    "@babel/plugin-transform-flow-strip-types": "7.0.0-beta.37",
-    "@babel/preset-react": "7.0.0-beta.37",
-    "babel-eslint": "8.2.1",
-    "body-parser": "1.18.2",
-    "create-universal-package": "2.1.1",
-    "eslint": "4.15.0",
-    "eslint-config-fusion": "0.2.1",
-    "eslint-plugin-cup": "1.0.0",
-    "eslint-plugin-flowtype": "2.41.0",
+    "@babel/preset-react": "7.0.0-beta.38",
+    "babel-eslint": "^8.2.1",
+    "body-parser": "^1.18.2",
+    "create-universal-package": "^3.2.5",
+    "eslint": "^4.15.0",
+    "eslint-config-fusion": "^0.2.1",
+    "eslint-plugin-cup": "^1.0.0",
+    "eslint-plugin-flowtype": "^2.41.0",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-prettier": "2.5.0",
-    "eslint-plugin-react": "7.5.1",
-    "express": "4.16.2",
-    "flow-bin": "0.63.1",
-    "fusion-core": "0.2.8",
-    "fusion-test-utils": "0.2.4",
-    "generic-session": "0.1.2",
-    "get-port": "3.2.0",
-    "nyc": "11.4.1",
+    "eslint-plugin-prettier": "^2.5.0",
+    "eslint-plugin-react": "^7.5.1",
+    "express": "^4.16.2",
+    "flow-bin": "^0.63.1",
+    "fusion-core": "^0.3.0-3",
+    "fusion-test-utils": "^0.4.0",
+    "fusion-tokens": "^0.0.4",
+    "generic-session": "^0.1.2",
+    "get-port": "^3.2.0",
+    "nyc": "^11.4.1",
     "prettier": "1.10.2",
-    "sinon": "4.1.5",
-    "tape-cup": "4.7.1",
-    "unitest": "2.1.1"
+    "sinon": "^4.1.6",
+    "tape-cup": "^4.7.1",
+    "unitest": "^2.1.1"
   },
   "peerDependencies": {
-    "fusion-core": "^0.2.3"
+    "fusion-core": "^0.3.0-3",
+    "fusion-tokens": "^0.0.4"
   },
   "scripts": {
     "clean": "rm -rf dist",
