+++ src/__tests__/index.node.js
@@ -1,11 +1,11 @@
 import App from 'fusion-core';
 import {request, render} from 'fusion-test-utils';
 import test from 'tape-cup';
-import CsrfPlugin from '../../server';
-import {GenericSessionToken} from 'fusion-types';
-import {request, render} from 'fusion-test-utils';
+import {GenericSessionToken} from 'fusion-tokens';
+import {getSimulator} from 'fusion-test-utils';
 import App from 'fusion-core';
-import {CSRFTokenExpire, CSRFIgnoreRoutes} from '../../shared';
+import CsrfPlugin from '../server';
+import {CSRFTokenExpire, CSRFIgnoreRoutes} from '../shared';
 
 function getSession() {
   const state = {};
@@ -27,24 +27,25 @@ function getSession() {
 test('valid token', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
-  const ctx = await request(app, '/csrf-token', {method: 'POST'});
+  const simulator = getSimulator(app);
+
+  const ctx = await simulator.request('/csrf-token', {method: 'POST'});
   t.ok(ctx.response.headers['x-csrf-token'], 'has token');
   t.equal(ctx.response.status, 200, 'has right status');
   t.equal(ctx.response.body, '', 'has empty body');
   const secret = Session.from(ctx).get('csrf-secret');
   t.ok(secret, 'sets a session secret');
 
-  const renderCtx = await render(app, '/');
+  const renderCtx = await simulator.render('/');
   t.equal(ctx.response.status, 200, 'has right status');
   t.equal(
     secret,
     Session.from(renderCtx).get('csrf-secret'),
     'does not change the session secret'
   );
-  const postCtx = await request(app, '/test', {
+  const postCtx = await simulator.request('/test', {
     method: 'POST',
     headers: {
       'x-csrf-token': ctx.response.headers['x-csrf-token'],
@@ -57,11 +58,11 @@ test('valid token', async t => {
 test('creates a session on a GET request', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
-  const ctx = await request(app, '/');
+  const ctx = await simulator.request('/');
   t.notok(
     ctx.response.headers['x-csrf-token'],
     'does not set x-csrf-token header'
@@ -74,11 +75,11 @@ test('creates a session on a GET request', async t => {
 test('render request', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
-  const ctx = await render(app, '/');
+  const ctx = await simulator.render('/');
   t.notok(
     ctx.response.headers['x-csrf-token'],
     'does not set x-csrf-token header'
@@ -95,12 +96,12 @@ test('render request', async t => {
 test('fails with no session and invalid token', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
   try {
-    await request(app, '/test', {
+    await simulator.request('/test', {
       method: 'POST',
       headers: {'x-csrf-token': 'invalid'},
     });
@@ -115,14 +116,14 @@ test('fails with no session and invalid token', async t => {
 test('fails with session and no token', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
-  await request(app, '/csrf-token', {method: 'POST'});
+  await simulator.request('/csrf-token', {method: 'POST'});
 
   try {
-    await request(app, '/test', {
+    await simulator.request('/test', {
       method: 'POST',
     });
     t.fail('should fail');
@@ -136,14 +137,14 @@ test('fails with session and no token', async t => {
 test('fails with session and invalid token', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
-  await request(app, '/csrf-token', {method: 'POST'});
+  await simulator.request('/csrf-token', {method: 'POST'});
 
   try {
-    await request(app, '/test', {
+    await simulator.request('/test', {
       method: 'POST',
       headers: {'x-csrf-token': 'invalid'},
     });
@@ -158,17 +159,17 @@ test('fails with session and invalid token', async t => {
 test('fails with expired token', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
-  app.configure(CSRFTokenExpire, 1);
+  app.register(GenericSessionToken, Session);
+  app.register(CSRFTokenExpire, 1);
   app.register(CsrfPlugin);
-  app.resolve();
+  const simulator = getSimulator(app);
 
-  const ctx = await request(app, '/csrf-token', {method: 'POST'});
+  const ctx = await simulator.request('/csrf-token', {method: 'POST'});
 
   await new Promise(resolve => setTimeout(resolve, 2000));
 
   try {
-    await request(app, '/test', {
+    await simulator.request('/test', {
       method: 'POST',
       headers: {'x-csrf-token': ctx.response.headers['x-csrf-token']},
     });
@@ -183,11 +184,11 @@ test('fails with expired token', async t => {
 test('does not verify ignored paths', async t => {
   const Session = getSession();
   const app = new App('fake-element', el => el);
-  app.register(GenericSessionToken, () => Session);
+  app.register(GenericSessionToken, Session);
   app.register(CsrfPlugin);
-  app.configure(CSRFIgnoreRoutes, ['/test']);
-  app.resolve();
-  const ctx = await request(app, '/test', {
+  app.register(CSRFIgnoreRoutes, ['/test']);
+  const simulator = getSimulator(app);
+  const ctx = await simulator.request('/test', {
     method: 'POST',
   });
   t.equal(ctx.response.status, 200);
