+++ src/__tests__/interface.js
@@ -2,6 +2,6 @@ import tape from 'tape-cup';
 import Plugin from '../index';
 
 tape('plugin api', t => {
-  t.equal(typeof Plugin, 'function', 'exports a default function');
+  t.ok(Plugin);
   t.end();
 });
